add sass file and linked to project
