+++ css/styles.css
@@ -200,7 +200,6 @@ h3 {
 
   textarea {
     height: 100px;
-    text
   }
 
   .save-btn {
@@ -235,7 +234,6 @@ h3 {
 
   textarea {
     height: 100px;
-    text
   }
 
   article {
