+++ lib/styles.scss
@@ -1,0 +1,5 @@
+// $primary-text-color: magenta;
+//
+// body {
+// background-color: $primary-text-color
+// }
