+++ test.html
@@ -5,6 +5,6 @@
   <title>2DoBox</title>
 </head>
 <body>
-  <script src="test.bundle.js"></script
+  <script src="test.bundle.js"></script>
 </body>
 </html>
