+++ webpack.config.js
@@ -1,5 +1,3 @@
-/ webpack.config.js
-
 const path = require('path');
 
 module.exports = {
@@ -15,7 +13,8 @@ module.exports = {
     loaders: [
       { test: /\.js$/, exclude: /node_modules/, loader: "babel-loader" },
       { test: /\.css$/, loader: "style!css" },
-      { test: /\.scss$/, loader: "style!css!sass" }
+      { test: /\.scss$/, loader: "style!css!sass" },
+      { test: /\.svg/, loader: "svg-url-loader" }
     ]
   },
   resolve: {
